[pax] Updated to patch 4.1.6-test17
Updated to pax-linux-4.1.6-test17.patch

PaX Team says:
  - Emese added global variable and structure field tracking to the size overflow plugin, see https://forums.grsecurity.net/viewtopic.php?f=7&t=3043&sid=f0c53309cf05a4c8c9ea7d5d516f4814#p15520 for details
    - there's an inevitable chance for triggering new bugs and false positives, please test carefully
  - Emese enabled the size overflow plugin on all archs but the hash tables represent x86 and arm only for now, feedback is welcome
  - cleaned up ktla_ktva/ktva_ktla on i386 to fix a gcc induced intentional integer overflow caught by the new overflow plugin
  - fixed some usb storage drivers to use SG_MAX to avoid triggering integer truncation on sg_tablesize assignments, caught by the new overflow plugin
  - fixed the limit setup code for the early percpu segment on i386, triggered by allyesconfig, reported by Emese
  - fixed a few compiler warnings about sign mismatches
  - fixed clashing definitions between struct irq_chip and irq_chip_no_const
  - fixed the atomic64_read_unchecked prototype on i386
  - fixed a few drivers for CONSTIFY/arm

+++ arch/x86/kernel/setup.c
@@ -782,7 +782,7 @@ static void __init trim_bios_range(void)
 /* called before trim_bios_range() to spare extra sanitize */
 static void __init e820_add_kernel_range(void)
 {
-	u64 start = __pa_symbol(ktla_ktva(_text));
+	u64 start = __pa_symbol(ktla_ktva((unsigned long)_text));
 	u64 size = __pa_symbol(_end) - start;
 
 	/*
@@ -962,15 +962,15 @@ void __init setup_arch(char **cmdline_p)
 
 	if (!boot_params.hdr.root_flags)
 		root_mountflags &= ~MS_RDONLY;
-	init_mm.start_code = ktla_ktva((unsigned long) _text);
-	init_mm.end_code = ktla_ktva((unsigned long) _etext);
-	init_mm.end_data = (unsigned long) _edata;
+	init_mm.start_code = ktla_ktva((unsigned long)_text);
+	init_mm.end_code = ktla_ktva((unsigned long)_etext);
+	init_mm.end_data = (unsigned long)_edata;
 	init_mm.brk = _brk_end;
 
 	mpx_mm_init(&init_mm);
 
-	code_resource.start = __pa_symbol(ktla_ktva(_text));
-	code_resource.end = __pa_symbol(ktla_ktva(_etext))-1;
+	code_resource.start = __pa_symbol(ktla_ktva((unsigned long)_text));
+	code_resource.end = __pa_symbol(ktla_ktva((unsigned long)_etext))-1;
 	data_resource.start = __pa_symbol(_sdata);
 	data_resource.end = __pa_symbol(_edata)-1;
 	bss_resource.start = __pa_symbol(__bss_start);
